Updated database path handling; Corrected sql query generation.
