+++ Makefile
@@ -1,0 +1,3 @@
+install:
+	swift build -c release
+	install .build/release/localizer /usr/local/bin/localizer
