+++ Sources/localizer/Catalog.swift
@@ -5,14 +5,15 @@ struct Catalog: ParsableCommand {
     
     static var configuration: CommandConfiguration = .init(
         commandName: "catalog",
-        abstract: "Interact with the catalog database.",
+        abstract: "Interact with the translation catalog.",
         discussion: """
         """,
         version: "1.0.0",
         shouldDisplay: true,
         subcommands: [
             Import.self,
-            Export.self
+            Export.self,
+            Generate.self
         ],
         defaultSubcommand: nil,
         helpNames: .shortAndLong
